requirements.txt
black==22.1.0
celery==5.2.3
dj-database-url==0.5.0
Django==3.2
django-allauth==0.48.0
django-celery-results==2.3.0
django-crispy-forms==1.14.0
django-rest-auth==0.9.5
djangorestframework==3.12.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
flake8==4.0.1
Pillow==9.0.1
postgres==4.0
psycopg2-binary==2.9.3
redis==4.1.4
swagger-spec-validator==2.7.4
waitress==2.1.1
whitenoise==6.0.0